requirements.txt
streamlit
paho-mqtt
streamlit-autorefresh